+++ crm/custom/Extension/application/Ext/Console/RegisterSugarAdminCliCommands.php
@@ -2,6 +2,10 @@
 
 use Sugarcrm\Sugarcrm\Console\CommandRegistry\CommandRegistry;
 use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\ClearAdditionalCacheCommand;
+use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\MaintenanceOffCommand;
+use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\MaintenanceOnCommand;
+use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\OrphansCleanupCommand;
+use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\PruneDatabaseCommand;
 use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\RebuildConfigCommand;
 use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\RebuildHtaccessCommand;
 use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\RebuildJsGroupingsCommand;
@@ -16,8 +20,10 @@ use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\RepairActiviti
 use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\RepairAndRebuildCommand;
 use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\RepairFieldCasingCommand;
 use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\RepairInboundEmailCommand;
+use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\RepairMissingTablesCommand;
 use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\RepairRolesCommand;
 use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\RepairTeamsCommand;
+use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\RestoreRecordCommand;
 use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\SeedUsersCommand;
 use Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command\UpgradeTeamsCommand;
 
@@ -42,4 +48,10 @@ CommandRegistry::getInstance()->addCommands([
     new RepairActivitiesCommand(),
     new SeedUsersCommand(),
     new ClearAdditionalCacheCommand(),
+    new RestoreRecordCommand(),
+    new OrphansCleanupCommand(),
+    new MaintenanceOnCommand(),
+    new MaintenanceOffCommand(),
+    new PruneDatabaseCommand(),
+    new RepairMissingTablesCommand(),
 ]);