+++ crm/custom/src/amaiza/SugarAdminCLI/Console/Command/MaintenanceOffCommand.php
@@ -1,0 +1,16 @@
+<?php
+namespace Sugarcrm\Sugarcrm\custom\amaiza\SugarAdminCLI\Console\Command;
+
+class MaintenanceOffCommand extends AbstractMaintenanceModeCommand {
+    protected function configure(): void
+    {
+        $this
+            ->setName('admin:maintenance:off')
+            ->setDescription('Set Maintenance mode off — the system becomes accessible to all users again.');
+    }
+
+    protected function targetStatus(): bool
+    {
+        return false;
+    }
+}